Move rest of dashboard code into routes.js and dashboard.js
- To reduce the size of index.js and serverTest.js
- To make it easier to find routes
- To make the test files more readable
- Also extended the tests for `POST /locations` to include checking
  whether the given values were actually stored in the DB
- Also fixed a `ALERT_TYPE.SENSOR_NO_MOTION` that should have been
  changed in the previous pull request
