+++ index.js
@@ -3,10 +3,7 @@ const bodyParser = require('body-parser')
 const express = require('express')
 const fs = require('fs')
 const https = require('https')
-const session = require('express-session')
-const cookieParser = require('cookie-parser')
 const cors = require('cors')
-const Mustache = require('mustache')
 const Validator = require('express-validator')
 
 // In-house dependencies
@@ -20,15 +17,7 @@ const BraveAlerterConfigurator = require('./BraveAlerterConfigurator')
 const IM21_DOOR_STATUS = require('./IM21DoorStatusEnum')
 const DOOR_STATUS = require('./SessionStateDoorEnum')
 const routes = require('./routes')
-
-const locationsDashboardTemplate = fs.readFileSync(`${__dirname}/mustache-templates/locationsDashboard.mst`, 'utf-8')
-const landingPageTemplate = fs.readFileSync(`${__dirname}/mustache-templates/landingPage.mst`, 'utf-8')
-const navPartial = fs.readFileSync(`${__dirname}/mustache-templates/navPartial.mst`, 'utf-8')
-const landingCSSPartial = fs.readFileSync(`${__dirname}/mustache-templates/landingCSSPartial.mst`, 'utf-8')
-const locationsCSSPartial = fs.readFileSync(`${__dirname}/mustache-templates/locationsCSSPartial.mst`, 'utf-8')
-const newLocationTemplate = fs.readFileSync(`${__dirname}/mustache-templates/newLocation.mst`, 'utf-8')
-const updateLocationTemplate = fs.readFileSync(`${__dirname}/mustache-templates/updateLocation.mst`, 'utf-8')
-const locationFormCSSPartial = fs.readFileSync(`${__dirname}/mustache-templates/locationFormCSSPartial.mst`, 'utf-8')
+const dashboard = require('./dashboard')
 
 // Start Express App
 const app = express()
@@ -47,64 +36,10 @@ app.use(bodyParser.json())
 // Cors Middleware (Cross Origin Resource Sharing)
 app.use(cors())
 
-app.use(cookieParser())
-
-// initialize express-session to allow us track the logged-in user across sessions.
-app.use(
-  session({
-    key: 'user_sid',
-    secret: helpers.getEnvVar('SECRET'),
-    resave: false,
-    saveUninitialized: false,
-    cookie: {
-      expires: 8 * 60 * 60 * 1000,
-    },
-  }),
-)
+dashboard.setupDashboardSessions(app)
 
 function convertToSeconds(milliseconds) {
   return Math.floor(milliseconds / 1000)
-}
-
-async function generateCalculatedTimeDifferenceString(timeToCompare) {
-  const daySecs = 24 * 60 * 60
-  const hourSecs = 60 * 60
-  const minSecs = 60
-  let returnString = ''
-  let numDays = 0
-  let numHours = 0
-  let numMins = 0
-  const currentTime = await db.getCurrentTime()
-
-  let diffSecs = (currentTime - timeToCompare) / 1000
-
-  if (diffSecs >= daySecs) {
-    numDays = Math.floor(diffSecs / daySecs)
-    diffSecs %= daySecs
-  }
-  returnString += `${numDays} ${numDays === 1 ? 'day, ' : 'days, '}`
-
-  if (diffSecs >= hourSecs) {
-    numHours = Math.floor(diffSecs / hourSecs)
-    diffSecs %= hourSecs
-  }
-  returnString += `${numHours} ${numHours === 1 ? 'hour, ' : 'hours, '}`
-
-  if (diffSecs >= minSecs) {
-    numMins = Math.floor(diffSecs / minSecs)
-  }
-  returnString += `${numMins} ${numMins === 1 ? 'minute' : 'minutes'}`
-
-  if (numDays + numHours === 0) {
-    diffSecs %= minSecs
-    const numSecs = Math.floor(diffSecs)
-
-    returnString += `, ${numSecs} ${numSecs === 1 ? 'second' : 'seconds'}`
-  }
-
-  returnString += ' ago'
-
-  return returnString
 }
 
 function getAlertTypeDisplayName(alertType) {
@@ -261,324 +196,6 @@ async function checkHeartbeat() {
     }
   }
 }
-
-// This middleware will check if user's cookie is still saved in browser and user is not set, then automatically log the user out.
-// This usually happens when you stop your express server after login, your cookie still remains saved in the browser.
-app.use((req, res, next) => {
-  if (req.cookies.user_sid && !req.session.user) {
-    res.clearCookie('user_sid')
-  }
-  next()
-})
-
-// middleware function to check for logged-in users
-function sessionChecker(req, res, next) {
-  if (!req.session.user || !req.cookies.user_sid) {
-    res.redirect('/login')
-  } else {
-    next()
-  }
-}
-
-app.get('/', sessionChecker, (req, res) => {
-  res.redirect('/dashboard')
-})
-
-app
-  .route('/login')
-  .get((req, res) => {
-    res.sendFile(`${__dirname}/login.html`)
-  })
-  .post((req, res) => {
-    const username = req.body.username
-    const password = req.body.password
-
-    if (username === helpers.getEnvVar('WEB_USERNAME') && password === helpers.getEnvVar('PASSWORD')) {
-      req.session.user = username
-      res.redirect('/dashboard')
-    } else {
-      res.redirect('/login')
-    }
-  })
-
-app.get('/logout', (req, res) => {
-  if (req.session.user && req.cookies.user_sid) {
-    res.clearCookie('user_sid')
-    res.redirect('/')
-  } else {
-    res.redirect('/login')
-  }
-})
-
-app.get('/dashboard', sessionChecker, async (req, res) => {
-  try {
-    // Needed for the navigation bar
-    const clients = await db.getClients()
-    const allLocations = await db.getLocations()
-
-    for (const location of allLocations) {
-      const recentSession = await db.getMostRecentSession(location.locationid)
-      if (recentSession !== null) {
-        const sessionCreatedAt = Date.parse(recentSession.createdAt)
-        const timeSinceLastSession = await generateCalculatedTimeDifferenceString(sessionCreatedAt)
-        location.sessionStart = timeSinceLastSession
-      }
-    }
-
-    const viewParams = {
-      clients,
-      locations: allLocations.map(location => {
-        return {
-          name: location.displayName,
-          id: location.locationid,
-          sessionStart: location.sessionStart,
-          isActive: location.isActive,
-          clientId: location.client.id,
-          clientDisplayName: location.client.displayName,
-        }
-      }),
-    }
-
-    res.send(Mustache.render(landingPageTemplate, viewParams, { nav: navPartial, css: landingCSSPartial }))
-  } catch (err) {
-    helpers.logError(`Error calling ${req.path}: ${err.toString()}`)
-    res.status(500).send()
-  }
-})
-
-app.get('/locations/new', sessionChecker, async (req, res) => {
-  try {
-    const clients = await db.getClients()
-    const viewParams = { clients }
-
-    res.send(Mustache.render(newLocationTemplate, viewParams, { nav: navPartial, css: locationFormCSSPartial }))
-  } catch (err) {
-    helpers.logError(`Error calling ${req.path}: ${err.toString()}`)
-    res.status(500).send()
-  }
-})
-
-app.get('/locations/:locationId', sessionChecker, async (req, res) => {
-  try {
-    // Needed for the navigation bar
-    const clients = await db.getClients()
-
-    const recentSessions = await db.getHistoryOfSessions(req.params.locationId)
-    const allLocations = await db.getLocations()
-    const currentLocation = allLocations.find(location => location.locationid === req.params.locationId)
-
-    const viewParams = {
-      clients,
-      recentSessions: [],
-      currentLocation,
-    }
-
-    // commented out keys were not shown on the old frontend but have been included in case that changes
-    for (const recentSession of recentSessions) {
-      const createdAt = recentSession.createdAt
-      const updatedAt = recentSession.updatedAt
-
-      viewParams.recentSessions.push({
-        createdAt,
-        updatedAt,
-        notes: recentSession.notes,
-        incidentType: recentSession.incidentType,
-        id: recentSession.id,
-        chatbotState: recentSession.chatbotState,
-        alertType: getAlertTypeDisplayName(recentSession.alertType),
-        respondedAt: recentSession.respondedAt,
-      })
-    }
-
-    res.send(Mustache.render(locationsDashboardTemplate, viewParams, { nav: navPartial, css: locationsCSSPartial }))
-  } catch (err) {
-    helpers.logError(`Error calling ${req.path}: ${err.toString()}`)
-    res.status(500).send()
-  }
-})
-
-app.get('/locations/:locationId/edit', sessionChecker, async (req, res) => {
-  try {
-    const clients = await db.getClients()
-    const allLocations = await db.getLocations()
-    const currentLocation = allLocations.find(location => location.locationid === req.params.locationId)
-
-    // for the dropdown in the edit screen so it does not display duplicate options
-    if (currentLocation.radarType === 'XeThru') {
-      currentLocation.otherType = 'Innosent'
-    } else {
-      currentLocation.otherType = 'XeThru'
-    }
-
-    const viewParams = {
-      currentLocation,
-      clients: clients.map(client => {
-        return {
-          ...client,
-          selected: client.id === currentLocation.client.id,
-        }
-      }),
-    }
-
-    res.send(Mustache.render(updateLocationTemplate, viewParams, { nav: navPartial, css: locationFormCSSPartial }))
-  } catch (err) {
-    helpers.logError(`Error calling ${req.path}: ${err.toString()}`)
-    res.status(500).send()
-  }
-})
-
-// Post routes for new add/update location forms
-
-app.post(
-  '/locations',
-  [
-    Validator.body([
-      'locationid',
-      'displayName',
-      'doorCoreID',
-      'radarCoreID',
-      'radarType',
-      'twilioPhone',
-      'firmwareStateMachine',
-      'clientId',
-    ]).notEmpty(),
-    Validator.oneOf([Validator.body(['responderPhoneNumber']).notEmpty(), Validator.body(['alertApiKey']).notEmpty()]),
-  ],
-  async (req, res) => {
-    try {
-      if (!req.session.user || !req.cookies.user_sid) {
-        helpers.logError('Unauthorized')
-        res.status(401).send()
-        return
-      }
-
-      const validationErrors = Validator.validationResult(req).formatWith(helpers.formatExpressValidationErrors)
-
-      if (validationErrors.isEmpty()) {
-        const allLocations = await db.getLocations()
-        const data = req.body
-
-        for (const location of allLocations) {
-          if (location.locationid === data.locationid) {
-            helpers.log('Location ID already exists')
-            return res.status(409).send('Location ID already exists')
-          }
-        }
-
-        const client = await db.getClientWithClientId(data.clientId)
-        if (client === null) {
-          const errorMessage = `Client ID '${data.clientId}' does not exist`
-          helpers.log(errorMessage)
-          return res.status(400).send(errorMessage)
-        }
-
-        await db.createLocationFromBrowserForm(
-          data.locationid,
-          data.displayName,
-          data.doorCoreID,
-          data.radarCoreID,
-          data.radarType,
-          data.responderPhoneNumber,
-          data.twilioPhone,
-          data.alertApiKey,
-          data.firmwareStateMachine,
-          data.clientId,
-        )
-
-        res.redirect(`/locations/${data.locationid}`)
-      } else {
-        const errorMessage = `Bad request to ${req.path}: ${validationErrors.array()}`
-        helpers.logError(errorMessage)
-        res.status(400).send(errorMessage)
-      }
-    } catch (err) {
-      helpers.logError(`Error calling ${req.path}: ${err.toString()}`)
-      res.status(500).send()
-    }
-  },
-)
-
-app.post(
-  '/locations/:locationId',
-  [
-    Validator.body([
-      'displayName',
-      'doorCoreID',
-      'radarCoreID',
-      'radarType',
-      'fallbackPhones',
-      'heartbeatPhones',
-      'twilioPhone',
-      'movementThreshold',
-      'durationTimer',
-      'stillnessTimer',
-      'initialTimer',
-      'reminderTimer',
-      'fallbackTimer',
-      'isActive',
-      'firmwareStateMachine',
-      'clientId',
-    ]).notEmpty(),
-    Validator.oneOf([Validator.body(['responderPhoneNumber']).notEmpty(), Validator.body(['alertApiKey']).notEmpty()]),
-  ],
-  async (req, res) => {
-    try {
-      if (!req.session.user || !req.cookies.user_sid) {
-        helpers.logError('Unauthorized')
-        res.status(401).send()
-        return
-      }
-
-      const validationErrors = Validator.validationResult(req).formatWith(helpers.formatExpressValidationErrors)
-
-      if (validationErrors.isEmpty()) {
-        const data = req.body
-        data.locationid = req.params.locationId
-
-        const client = await db.getClientWithClientId(data.clientId)
-        if (client === null) {
-          const errorMessage = `Client ID '${data.clientId}' does not exist`
-          helpers.log(errorMessage)
-          return res.status(400).send(errorMessage)
-        }
-
-        const newAlertApiKey = data.alertApiKey && data.alertApiKey.trim() !== '' ? data.alertApiKey : null
-        const newPhone = data.responderPhoneNumber && data.responderPhoneNumber.trim() !== '' ? data.responderPhoneNumber : null
-
-        await db.updateLocation(
-          data.displayName,
-          data.doorCoreID,
-          data.radarCoreID,
-          data.radarType,
-          newPhone,
-          data.fallbackPhones.split(','),
-          data.heartbeatPhones.split(','),
-          data.twilioPhone,
-          data.movementThreshold,
-          data.durationTimer,
-          data.stillnessTimer,
-          data.initialTimer,
-          data.reminderTimer,
-          data.fallbackTimer,
-          newAlertApiKey,
-          data.isActive === 'true',
-          data.firmwareStateMachine === 'true',
-          data.locationid,
-          data.clientId,
-        )
-
-        res.redirect(`/locations/${data.locationid}`)
-      } else {
-        const errorMessage = `Bad request to ${req.path}: ${validationErrors.array()}`
-        helpers.logError(errorMessage)
-        res.status(400).send(errorMessage)
-      }
-    } catch (err) {
-      helpers.logError(`Error calling ${req.path}: ${err.toString()}`)
-      res.status(500).send()
-    }
-  },
-)
 
 // Add routes
 routes.configureRoutes(app)
