+++ mustache-templates/clientPage.mst
@@ -22,7 +22,7 @@
 
                 <a href="/clients/{{id}}/edit" class="btn btn-secondary btn-sm" role="button">Edit Client</a>
                 <br>
-
+                <br>
                 <div class="table-responsive">
                     <table class="table table-striped table-sm table-fixed">
                         <thead>
