+++ test/integration/dashboardTest/submitNewClientTest.js
@@ -24,6 +24,8 @@ describe('dashboard.js integration tests: submitNewClient', () => {
     sandbox.spy(helpers, 'log')
     sandbox.spy(helpers, 'logError')
 
+    await db.clearSessions()
+    await db.clearLocations()
     await db.clearClients()
 
     this.agent = chai.request.agent(server)
@@ -31,7 +33,11 @@ describe('dashboard.js integration tests: submitNewClient', () => {
 
   afterEach(async () => {
     this.agent.close()
+
+    await db.clearSessions()
+    await db.clearLocations()
     await db.clearClients()
+
     sandbox.restore()
   })
