+++ test/integration/serverTest.js
@@ -822,857 +822,5 @@ describe('Brave Sensor server', () => {
         expect(response).to.have.status(400)
       })
     })
-
-    describe('POST /locations', () => {
-      describe('for a request that contains valid non-empty fields', () => {
-        beforeEach(async () => {
-          await redis.clearKeys()
-          await db.clearSessions()
-          await db.clearLocations()
-          await db.clearClients()
-
-          this.client = await clientFactory(db)
-
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          const goodRequest = {
-            locationid: 'unusedID',
-            displayName: 'locationName',
-            doorCoreID: door_coreID,
-            radarCoreID: radar_coreID,
-            radarType: 'XeThru',
-            responderPhoneNumber: testLocation1PhoneNumber,
-            twilioPhone: '+15005550006',
-            firmwareStateMachine: 'false',
-            clientId: this.client.id,
-          }
-
-          this.response = await this.agent.post('/locations').send(goodRequest)
-        })
-
-        afterEach(async () => {
-          await redis.clearKeys()
-          await db.clearSessions()
-          await db.clearLocations()
-          await db.clearClients()
-
-          this.agent.close()
-        })
-
-        it('should return 200', () => {
-          expect(this.response).to.have.status(200)
-        })
-
-        it('should create a location in the database', async () => {
-          const newLocation = await db.getLocationData('unusedID')
-
-          expect(newLocation).to.not.be.undefined
-        })
-      })
-
-      describe('for a request with an non-existant client ID', () => {
-        beforeEach(async () => {
-          sinon.stub(helpers, 'log')
-          sinon.spy(db, 'createLocationFromBrowserForm')
-
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          this.clientId = 'fbb3e19d-5884-46d6-a0e5-ed5c7c406274'
-          const goodRequest = {
-            locationid: 'unusedID',
-            displayName: 'locationName',
-            doorCoreID: door_coreID,
-            radarCoreID: radar_coreID,
-            radarType: 'XeThru',
-            responderPhoneNumber: testLocation1PhoneNumber,
-            twilioPhone: '+15005550006',
-            firmwareStateMachine: 'false',
-            clientId: this.clientId,
-          }
-
-          this.response = await this.agent.post('/locations').send(goodRequest)
-        })
-
-        afterEach(async () => {
-          helpers.log.restore()
-          db.createLocationFromBrowserForm.restore()
-
-          this.agent.close()
-        })
-
-        it('should return 400', () => {
-          expect(this.response).to.have.status(400)
-        })
-
-        it('should not create a new location in the database', () => {
-          expect(db.createLocationFromBrowserForm).to.not.have.been.called
-        })
-
-        it('should log the error', () => {
-          expect(helpers.log).to.have.been.calledWith(`Client ID '${this.clientId}' does not exist`)
-        })
-      })
-
-      describe('for a request with no login session', () => {
-        beforeEach(async () => {
-          sinon.stub(helpers, 'logError')
-          sinon.spy(db, 'createLocationFromBrowserForm')
-
-          const goodRequest = {
-            locationid: 'unusedID',
-            displayName: 'locationName',
-            doorCoreID: door_coreID,
-            radarCoreID: radar_coreID,
-            radarType: 'XeThru',
-            responderPhoneNumber: testLocation1PhoneNumber,
-            twilioPhone: '+15005550006',
-            alertApiKey: 'myApiKey',
-            firmwareStateMachine: 'false',
-            clientId: '91ddc8f7-c2e7-490e-bfe9-3d2880a76108',
-          }
-
-          this.response = await chai.request(server).post('/locations').send(goodRequest)
-        })
-
-        afterEach(async () => {
-          helpers.logError.restore()
-          db.createLocationFromBrowserForm.restore()
-        })
-
-        it('should return 401', () => {
-          expect(this.response).to.have.status(401)
-        })
-
-        it('should not create a new location in the database', () => {
-          expect(db.createLocationFromBrowserForm).to.not.have.been.called
-        })
-
-        it('should log the error', () => {
-          expect(helpers.logError).to.have.been.calledWith('Unauthorized')
-        })
-      })
-
-      describe('for a request that contains all valid non-empty fields but with no alertApiKey', () => {
-        beforeEach(async () => {
-          await redis.clearKeys()
-          await db.clearSessions()
-          await db.clearLocations()
-          await db.clearClients()
-
-          this.client = await clientFactory(db)
-
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          const goodRequest = {
-            locationid: 'unusedID',
-            displayName: 'locationName',
-            doorCoreID: door_coreID,
-            radarCoreID: radar_coreID,
-            radarType: 'XeThru',
-            responderPhoneNumber: testLocation1PhoneNumber,
-            twilioPhone: '+15005550006',
-            firmwareStateMachine: 'false',
-            clientId: this.client.id,
-          }
-
-          this.response = await this.agent.post('/locations').send(goodRequest)
-        })
-
-        afterEach(async () => {
-          await redis.clearKeys()
-          await db.clearSessions()
-          await db.clearLocations()
-          await db.clearClients()
-
-          this.agent.close()
-        })
-
-        it('should return 200', () => {
-          expect(this.response).to.have.status(200)
-        })
-
-        it('should create a location in the database', async () => {
-          const newLocation = await db.getLocationData('unusedID')
-
-          expect(newLocation).to.not.be.undefined
-        })
-      })
-
-      describe('for a request that contains all valid non-empty fields but with no phone', () => {
-        beforeEach(async () => {
-          await redis.clearKeys()
-          await db.clearSessions()
-          await db.clearLocations()
-          await db.clearClients()
-
-          this.client = await clientFactory(db)
-
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          const goodRequest = {
-            locationid: 'unusedID',
-            displayName: 'locationName',
-            doorCoreID: door_coreID,
-            radarCoreID: radar_coreID,
-            radarType: 'XeThru',
-            twilioPhone: '+15005550006',
-            alertApiKey: 'alertApiKey',
-            firmwareStateMachine: 'false',
-            clientId: this.client.id,
-          }
-
-          this.response = await this.agent.post('/locations').send(goodRequest)
-        })
-
-        afterEach(async () => {
-          await redis.clearKeys()
-          await db.clearSessions()
-          await db.clearLocations()
-          await db.clearClients()
-
-          this.agent.close()
-        })
-
-        it('should return 200', () => {
-          expect(this.response).to.have.status(200)
-        })
-
-        it('should create a location in the database', async () => {
-          const newLocation = await db.getLocationData('unusedID')
-
-          expect(newLocation).to.not.be.undefined
-        })
-      })
-
-      describe('for a request that contains all valid fields, but empty', () => {
-        beforeEach(async () => {
-          sinon.stub(helpers, 'logError')
-          sinon.spy(db, 'createLocationFromBrowserForm')
-
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          const badRequest = {
-            locationid: '',
-            displayName: '',
-            doorCoreID: '',
-            radarCoreID: '',
-            radarType: '',
-            responderPhoneNumber: '',
-            twilioPhone: '',
-            alertApiKey: '',
-            firmwareStateMachine: '',
-            clientId: '',
-          }
-
-          this.response = await this.agent.post('/locations').send(badRequest)
-        })
-
-        afterEach(async () => {
-          helpers.logError.restore()
-          db.createLocationFromBrowserForm.restore()
-
-          this.agent.close()
-        })
-
-        it('should return 400', () => {
-          expect(this.response).to.have.status(400)
-        })
-
-        it('should not create a new location in the database', () => {
-          expect(db.createLocationFromBrowserForm).to.not.have.been.called
-        })
-
-        it('should log the error', () => {
-          expect(helpers.logError).to.have.been.calledWith(
-            `Bad request to /locations: locationid (Invalid value),displayName (Invalid value),doorCoreID (Invalid value),radarCoreID (Invalid value),radarType (Invalid value),twilioPhone (Invalid value),firmwareStateMachine (Invalid value),clientId (Invalid value),responderPhoneNumber/alertApiKey (Invalid value(s))`,
-          )
-        })
-      })
-
-      describe('for an empty request', () => {
-        beforeEach(async () => {
-          sinon.stub(helpers, 'logError')
-          sinon.spy(db, 'createLocationFromBrowserForm')
-
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          this.response = await this.agent.post('/locations').send({})
-        })
-
-        afterEach(() => {
-          helpers.logError.restore()
-          db.createLocationFromBrowserForm.restore()
-
-          this.agent.close()
-        })
-
-        it('should return 400', () => {
-          expect(this.response).to.have.status(400)
-        })
-
-        it('should not create a new location in the database', () => {
-          expect(db.createLocationFromBrowserForm).to.not.have.been.called
-        })
-
-        it('should log the error', () => {
-          expect(helpers.logError).to.have.been.calledWith(
-            `Bad request to /locations: locationid (Invalid value),displayName (Invalid value),doorCoreID (Invalid value),radarCoreID (Invalid value),radarType (Invalid value),twilioPhone (Invalid value),firmwareStateMachine (Invalid value),clientId (Invalid value),responderPhoneNumber/alertApiKey (Invalid value(s))`,
-          )
-        })
-      })
-
-      describe('for an otherwise valid request that contains an already existing locationid', () => {
-        beforeEach(async () => {
-          sinon.stub(helpers, 'log')
-          sinon.spy(db, 'createLocationFromBrowserForm')
-
-          this.client = await clientFactory(db)
-
-          await db.clearLocations()
-          await db.createLocation(
-            testLocation1Id,
-            testLocation1PhoneNumber,
-            MOVEMENT_THRESHOLD,
-            15,
-            1,
-            5000,
-            5000,
-            ['+15005550006'],
-            '+15005550006',
-            ['+15005550006'],
-            1000,
-            'locationName',
-            door_coreID,
-            radar_coreID,
-            'XeThru',
-            'alertApiKey',
-            true,
-            false,
-            this.client.id,
-          )
-
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          const duplicateLocationRequest = {
-            locationid: testLocation1Id,
-            displayName: 'locationName',
-            doorCoreID: door_coreID,
-            radarCoreID: radar_coreID,
-            radarType: 'XeThru',
-            responderPhoneNumber: testLocation1PhoneNumber,
-            twilioPhone: '+15005550006',
-            alertApiKey: 'alertApiKey',
-            firmwareStateMachine: 'false',
-            clientId: this.client.id,
-          }
-
-          this.response = await this.agent.post('/locations').send(duplicateLocationRequest)
-        })
-
-        afterEach(async () => {
-          helpers.log.restore()
-          db.createLocationFromBrowserForm.restore()
-
-          this.agent.close()
-
-          await db.clearLocations()
-          await db.clearClients()
-        })
-
-        it('should return 409', () => {
-          expect(this.response).to.have.status(409)
-        })
-
-        it('should not create a new location in the database', () => {
-          expect(db.createLocationFromBrowserForm).to.not.have.been.called
-        })
-
-        it('should log the error', () => {
-          expect(helpers.log).to.have.been.calledWith('Location ID already exists')
-        })
-      })
-    })
-
-    describe('POST /locations/:locationId', () => {
-      beforeEach(async () => {
-        sinon.spy(helpers, 'log')
-        sinon.spy(helpers, 'logError')
-        sinon.spy(db, 'updateLocation')
-
-        this.testLocationIdForEdit = 'test1'
-        await db.clearLocations()
-        await db.clearClients()
-
-        this.client = await clientFactory(db)
-        await db.createLocation(
-          this.testLocationIdForEdit,
-          '+14445556789',
-          10,
-          50,
-          100,
-          90000,
-          15000,
-          ['+14445556789'],
-          '+14445556789',
-          ['+14445556789'],
-          90000,
-          'Initial Name',
-          'door_core',
-          'radar_core',
-          'XeThru',
-          'alertApiKey',
-          true,
-          false,
-          this.client.id,
-        )
-      })
-
-      afterEach(async () => {
-        helpers.log.restore()
-        helpers.logError.restore()
-        db.updateLocation.restore()
-
-        await db.clearLocations()
-        await db.clearClients()
-      })
-
-      describe('for a request that contains all valid non-empty fields', () => {
-        beforeEach(async () => {
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          this.goodRequest = {
-            displayName: 'New Name',
-            doorCoreID: 'new_door_core',
-            radarCoreID: 'new_radar_core',
-            radarType: 'Innosent',
-            responderPhoneNumber: '+12223334567',
-            fallbackPhones: '+12223334444,+13334445678',
-            heartbeatPhones: '+15556667890,+16667778901',
-            twilioPhone: '+11112223456',
-            movementThreshold: 15,
-            durationTimer: 90,
-            stillnessTimer: 10,
-            initialTimer: 9856,
-            reminderTimer: 567849,
-            fallbackTimer: 234567,
-            alertApiKey: 'newApiKey',
-            isActive: 'true',
-            firmwareStateMachine: 'false',
-            clientId: this.client.id,
-          }
-
-          this.response = await this.agent.post(`/locations/${this.testLocationIdForEdit}`).send(this.goodRequest)
-        })
-
-        afterEach(() => {
-          this.agent.close()
-        })
-
-        it('should return 200', () => {
-          expect(this.response).to.have.status(200)
-        })
-
-        it('should update the location in the database', async () => {
-          const updatedLocation = await db.getLocationData(this.testLocationIdForEdit)
-
-          expect(updatedLocation.displayName).to.equal(this.goodRequest.displayName)
-          expect(updatedLocation.doorCoreId).to.equal(this.goodRequest.doorCoreID)
-          expect(updatedLocation.radarCoreId).to.equal(this.goodRequest.radarCoreID)
-          expect(updatedLocation.radarType).to.equal(this.goodRequest.radarType)
-          expect(updatedLocation.responderPhoneNumber).to.equal(this.goodRequest.responderPhoneNumber)
-          expect(updatedLocation.fallbackNumbers.join(',')).to.equal(this.goodRequest.fallbackPhones)
-          expect(updatedLocation.heartbeatAlertRecipients.join(',')).to.equal(this.goodRequest.heartbeatPhones)
-          expect(updatedLocation.twilioNumber).to.equal(this.goodRequest.twilioPhone)
-          expect(updatedLocation.alertApiKey).to.equal(this.goodRequest.alertApiKey)
-          expect(updatedLocation.isActive).to.be.true
-          expect(updatedLocation.firmwareStateMachine).to.be.false
-          expect(updatedLocation.client.id).to.equal(this.goodRequest.clientId)
-
-          chai.assert.equal(updatedLocation.movementThreshold, this.goodRequest.movementThreshold)
-          chai.assert.equal(updatedLocation.durationTimer, this.goodRequest.durationTimer)
-          chai.assert.equal(updatedLocation.stillnessTimer, this.goodRequest.stillnessTimer)
-          chai.assert.equal(updatedLocation.initialTimer, this.goodRequest.initialTimer)
-          chai.assert.equal(updatedLocation.reminderTimer, this.goodRequest.reminderTimer)
-          chai.assert.equal(updatedLocation.fallbackTimer, this.goodRequest.fallbackTimer)
-        })
-      })
-
-      describe('for a request that contains all valid non-empty fields but is inactive', () => {
-        beforeEach(async () => {
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          this.goodRequest = {
-            displayName: 'New Name',
-            doorCoreID: 'new_door_core',
-            radarCoreID: 'new_radar_core',
-            radarType: 'Innosent',
-            responderPhoneNumber: '+12223334567',
-            fallbackPhones: '+12223334444,+13334445678',
-            heartbeatPhones: '+15556667890,+16667778901',
-            twilioPhone: '+11112223456',
-            movementThreshold: 15,
-            durationTimer: 90,
-            stillnessTimer: 10,
-            initialTimer: 9856,
-            reminderTimer: 567849,
-            fallbackTimer: 234567,
-            alertApiKey: 'newApiKey',
-            isActive: 'false',
-            firmwareStateMachine: 'false',
-            clientId: this.client.id,
-          }
-
-          this.response = await this.agent.post(`/locations/${this.testLocationIdForEdit}`).send(this.goodRequest)
-        })
-
-        afterEach(() => {
-          this.agent.close()
-        })
-
-        it('should return 200', () => {
-          expect(this.response).to.have.status(200)
-        })
-
-        it('should update the location in the database', async () => {
-          const updatedLocation = await db.getLocationData(this.testLocationIdForEdit)
-
-          expect(updatedLocation.displayName).to.equal(this.goodRequest.displayName)
-          expect(updatedLocation.doorCoreId).to.equal(this.goodRequest.doorCoreID)
-          expect(updatedLocation.radarCoreId).to.equal(this.goodRequest.radarCoreID)
-          expect(updatedLocation.radarType).to.equal(this.goodRequest.radarType)
-          expect(updatedLocation.responderPhoneNumber).to.equal(this.goodRequest.responderPhoneNumber)
-          expect(updatedLocation.fallbackNumbers.join(',')).to.equal(this.goodRequest.fallbackPhones)
-          expect(updatedLocation.heartbeatAlertRecipients.join(',')).to.equal(this.goodRequest.heartbeatPhones)
-          expect(updatedLocation.twilioNumber).to.equal(this.goodRequest.twilioPhone)
-          expect(updatedLocation.alertApiKey).to.equal(this.goodRequest.alertApiKey)
-          expect(updatedLocation.isActive).to.be.false
-          expect(updatedLocation.firmwareStateMachine).to.be.false
-          expect(updatedLocation.client.id).to.equal(this.goodRequest.clientId)
-
-          chai.assert.equal(updatedLocation.movementThreshold, this.goodRequest.movementThreshold)
-          chai.assert.equal(updatedLocation.durationTimer, this.goodRequest.durationTimer)
-          chai.assert.equal(updatedLocation.stillnessTimer, this.goodRequest.stillnessTimer)
-          chai.assert.equal(updatedLocation.initialTimer, this.goodRequest.initialTimer)
-          chai.assert.equal(updatedLocation.reminderTimer, this.goodRequest.reminderTimer)
-          chai.assert.equal(updatedLocation.fallbackTimer, this.goodRequest.fallbackTimer)
-        })
-      })
-
-      describe('for a request that contains all valid non-empty fields but with an empty alertApiKey', () => {
-        beforeEach(async () => {
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          this.goodRequest = {
-            displayName: 'New Name',
-            doorCoreID: 'new_door_core',
-            radarCoreID: 'new_radar_core',
-            radarType: 'Innosent',
-            responderPhoneNumber: '+12223334567',
-            fallbackPhones: '+13334445678,+12223334444',
-            heartbeatPhones: '+15556667890,+16667778901',
-            twilioPhone: '+11112223456',
-            movementThreshold: 15,
-            durationTimer: 90,
-            stillnessTimer: 10,
-            initialTimer: 9856,
-            reminderTimer: 567849,
-            fallbackTimer: 234567,
-            alertApiKey: '',
-            isActive: 'true',
-            firmwareStateMachine: 'false',
-            clientId: this.client.id,
-          }
-
-          this.response = await this.agent.post(`/locations/${this.testLocationIdForEdit}`).send(this.goodRequest)
-        })
-
-        afterEach(() => {
-          this.agent.close()
-        })
-
-        it('should return 200', () => {
-          expect(this.response).to.have.status(200)
-        })
-
-        it('should update the location in the database', async () => {
-          const updatedLocation = await db.getLocationData(this.testLocationIdForEdit)
-
-          expect(updatedLocation.displayName).to.equal(this.goodRequest.displayName)
-          expect(updatedLocation.doorCoreId).to.equal(this.goodRequest.doorCoreID)
-          expect(updatedLocation.radarCoreId).to.equal(this.goodRequest.radarCoreID)
-          expect(updatedLocation.radarType).to.equal(this.goodRequest.radarType)
-          expect(updatedLocation.responderPhoneNumber).to.equal(this.goodRequest.responderPhoneNumber)
-          expect(updatedLocation.fallbackNumbers.join(',')).to.equal(this.goodRequest.fallbackPhones)
-          expect(updatedLocation.heartbeatAlertRecipients.join(',')).to.equal(this.goodRequest.heartbeatPhones)
-          expect(updatedLocation.twilioNumber).to.equal(this.goodRequest.twilioPhone)
-          expect(updatedLocation.alertApiKey).to.be.null
-          expect(updatedLocation.isActive).to.be.true
-          expect(updatedLocation.firmwareStateMachine).to.be.false
-          expect(updatedLocation.client.id).to.equal(this.goodRequest.clientId)
-
-          chai.assert.equal(updatedLocation.movementThreshold, this.goodRequest.movementThreshold)
-          chai.assert.equal(updatedLocation.durationTimer, this.goodRequest.durationTimer)
-          chai.assert.equal(updatedLocation.stillnessTimer, this.goodRequest.stillnessTimer)
-          chai.assert.equal(updatedLocation.initialTimer, this.goodRequest.initialTimer)
-          chai.assert.equal(updatedLocation.reminderTimer, this.goodRequest.reminderTimer)
-          chai.assert.equal(updatedLocation.fallbackTimer, this.goodRequest.fallbackTimer)
-        })
-      })
-
-      describe('for a request that contains all valid non-empty fields but with an empty phone', () => {
-        beforeEach(async () => {
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          this.goodRequest = {
-            displayName: 'New Name',
-            doorCoreID: 'new_door_core',
-            radarCoreID: 'new_radar_core',
-            radarType: 'Innosent',
-            responderPhoneNumber: '',
-            fallbackPhones: '+12223334444,+13334445678',
-            heartbeatPhones: '+15556667890,+16667778901',
-            twilioPhone: '+11112223456',
-            movementThreshold: 15,
-            durationTimer: 90,
-            stillnessTimer: 10,
-            initialTimer: 9856,
-            reminderTimer: 567849,
-            fallbackTimer: 234567,
-            alertApiKey: 'newApiKey',
-            isActive: 'true',
-            firmwareStateMachine: 'false',
-            clientId: this.client.id,
-          }
-
-          this.response = await this.agent.post(`/locations/${this.testLocationIdForEdit}`).send(this.goodRequest)
-        })
-
-        afterEach(() => {
-          this.agent.close()
-        })
-
-        it('should return 200', () => {
-          expect(this.response).to.have.status(200)
-        })
-
-        it('should update the location in the database', async () => {
-          const updatedLocation = await db.getLocationData(this.testLocationIdForEdit)
-
-          expect(updatedLocation.displayName).to.equal(this.goodRequest.displayName)
-          expect(updatedLocation.doorCoreId).to.equal(this.goodRequest.doorCoreID)
-          expect(updatedLocation.radarCoreId).to.equal(this.goodRequest.radarCoreID)
-          expect(updatedLocation.radarType).to.equal(this.goodRequest.radarType)
-          expect(updatedLocation.responderPhoneNumber).to.be.null
-          expect(updatedLocation.fallbackNumbers.join(',')).to.equal(this.goodRequest.fallbackPhones)
-          expect(updatedLocation.heartbeatAlertRecipients.join(',')).to.equal(this.goodRequest.heartbeatPhones)
-          expect(updatedLocation.twilioNumber).to.equal(this.goodRequest.twilioPhone)
-          expect(updatedLocation.alertApiKey).to.equal(this.goodRequest.alertApiKey)
-          expect(updatedLocation.isActive).to.be.true
-          expect(updatedLocation.firmwareStateMachine).to.be.false
-          expect(updatedLocation.client.id).to.equal(this.goodRequest.clientId)
-
-          chai.assert.equal(updatedLocation.movementThreshold, this.goodRequest.movementThreshold)
-          chai.assert.equal(updatedLocation.durationTimer, this.goodRequest.durationTimer)
-          chai.assert.equal(updatedLocation.stillnessTimer, this.goodRequest.stillnessTimer)
-          chai.assert.equal(updatedLocation.initialTimer, this.goodRequest.initialTimer)
-          chai.assert.equal(updatedLocation.reminderTimer, this.goodRequest.reminderTimer)
-          chai.assert.equal(updatedLocation.fallbackTimer, this.goodRequest.fallbackTimer)
-        })
-      })
-
-      describe('for a request that contains all valid non-empty fields but with a non-existant clientId', () => {
-        beforeEach(async () => {
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          this.clientId = '8244c552-6753-4713-bbb6-07ad1c7fb8f8'
-          this.goodRequest = {
-            displayName: 'New Name',
-            doorCoreID: 'new_door_core',
-            radarCoreID: 'new_radar_core',
-            radarType: 'Innosent',
-            responderPhoneNumber: '+12223334567',
-            fallbackPhones: '+12223334444,+13334445678',
-            heartbeatPhones: '+15556667890,+16667778901',
-            twilioPhone: '+11112223456',
-            movementThreshold: 15,
-            durationTimer: 90,
-            stillnessTimer: 10,
-            initialTimer: 9856,
-            reminderTimer: 567849,
-            fallbackTimer: 234567,
-            alertApiKey: 'newApiKey',
-            isActive: 'true',
-            firmwareStateMachine: 'false',
-            clientId: this.clientId,
-          }
-
-          this.response = await this.agent.post(`/locations/${this.testLocationIdForEdit}`).send(this.goodRequest)
-        })
-
-        afterEach(() => {
-          this.agent.close()
-        })
-
-        it('should return 400', () => {
-          expect(this.response).to.have.status(400)
-        })
-
-        it('should not update the location in the database', () => {
-          expect(db.updateLocation).to.not.have.been.called
-        })
-
-        it('should log the error', () => {
-          expect(helpers.log).to.have.been.calledWith(`Client ID '${this.clientId}' does not exist`)
-        })
-      })
-
-      describe('for a request that contains all valid fields, but empty', () => {
-        beforeEach(async () => {
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          // We are not testing any logs that happen because of the setup
-          helpers.log.resetHistory()
-
-          const badRequest = {
-            displayName: '',
-            doorCoreID: '',
-            radarCoreID: '',
-            radarType: '',
-            responderPhoneNumber: '',
-            fallbackPhones: '',
-            heartbeatPhones: '',
-            twilioPhone: '',
-            movementThreshold: '',
-            durationTimer: '',
-            stillnessTimer: '',
-            initialTimer: '',
-            reminderTimer: '',
-            fallbackTimer: '',
-            alertApiKey: '',
-            isActive: '',
-            clientId: '',
-          }
-
-          this.response = await this.agent.post(`/locations/${testLocation1Id}`).send(badRequest)
-        })
-
-        afterEach(() => {
-          this.agent.close()
-        })
-
-        it('should return 400', () => {
-          expect(this.response).to.have.status(400)
-        })
-
-        it('should not update the location in the database', () => {
-          expect(db.updateLocation).to.not.have.been.called
-        })
-
-        it('should log the error', () => {
-          expect(helpers.logError).to.have.been.calledWith(
-            `Bad request to /locations/TestLocation1: displayName (Invalid value),doorCoreID (Invalid value),radarCoreID (Invalid value),radarType (Invalid value),fallbackPhones (Invalid value),heartbeatPhones (Invalid value),twilioPhone (Invalid value),movementThreshold (Invalid value),durationTimer (Invalid value),stillnessTimer (Invalid value),initialTimer (Invalid value),reminderTimer (Invalid value),fallbackTimer (Invalid value),isActive (Invalid value),firmwareStateMachine (Invalid value),clientId (Invalid value),responderPhoneNumber/alertApiKey (Invalid value(s))`,
-          )
-        })
-      })
-
-      describe('for an empty request', () => {
-        beforeEach(async () => {
-          this.agent = chai.request.agent(server)
-
-          await this.agent.post('/login').send({
-            username: helpers.getEnvVar('WEB_USERNAME'),
-            password: helpers.getEnvVar('PASSWORD'),
-          })
-
-          // We are not testing any logs that happen because of the setup
-          helpers.log.resetHistory()
-
-          this.response = await this.agent.post(`/locations/${testLocation1Id}`).send({})
-        })
-
-        afterEach(() => {
-          this.agent.close()
-        })
-
-        it('should return 400', () => {
-          expect(this.response).to.have.status(400)
-        })
-
-        it('should not update the location in the database', () => {
-          expect(db.updateLocation).to.not.have.been.called
-        })
-
-        it('should log the error', () => {
-          expect(helpers.logError).to.have.been.calledWith(
-            `Bad request to /locations/TestLocation1: displayName (Invalid value),doorCoreID (Invalid value),radarCoreID (Invalid value),radarType (Invalid value),fallbackPhones (Invalid value),heartbeatPhones (Invalid value),twilioPhone (Invalid value),movementThreshold (Invalid value),durationTimer (Invalid value),stillnessTimer (Invalid value),initialTimer (Invalid value),reminderTimer (Invalid value),fallbackTimer (Invalid value),isActive (Invalid value),firmwareStateMachine (Invalid value),clientId (Invalid value),responderPhoneNumber/alertApiKey (Invalid value(s))`,
-          )
-        })
-      })
-    })
   })
 })
